initial changes to support cuda 12.8
In this change we do a few things:
- We add support for CUDA 12.8 in the install.sh script.
- We add support for CUDA 12.8 in the requirements-uv.txt and requirements.in
files using the new `--index` flag. to match the instructions for installing
pytorch according to the pytorch website.
- There is a strange change: we had to stop using $HOST as an env
variable in the run.sh script. This is because it was causing issues with
the conda activate -- somehow something is setting HOST to different value
than the one we are using in the script. So we just change the variable name
- When running plugins, they activate a venv per plugin, this activation
scripts need to know to add the extra index or they won't be able to find
pytorch 2.7 with cuda 12.8
- NOTDONE: We should upgrade all plugin versions to force a rebuild of each venv
- Also note: a clean upgrade of the same conda env didn't work for me,
I had to wipe it and start again. It might work for you.